chore: move translation status generation to script
