--- docs/locales.d.ts
@@ -1,11 +1,0 @@
-declare module 'virtual:elk-locales' {
-    const locales: Record<string, {
-        title: string
-        translated: string[]
-        missing: string[]
-        outdated: string[]
-        total: number
-        isSource?: boolean
-    }>
-    export default locales
-}
